[STRUCTURAL] Move Method: orphan sealing to pkg/agentsdk
The walk that seals unanswered tool_use blocks existed only in
internal/agent, so the portable loop had no way to reach it. Move it to
pkg/agentsdk.SealOrphanedToolUses over a small ToolResultConversation
interface — Messages() plus AddToolResult — which both packages'
Conversation types already satisfy, since internal/agent.Message is an
alias of agentsdk.Message.

internal/agent.synthesizeMissingToolResults now delegates; its reason
constants and every call site are unchanged, so behavior is identical and
the existing orphan suite covers it. The point is that the SDK loop can
now call the same implementation rather than growing a second copy.

Refs #325.
